Add graph filter

+++ app/src/main/java/com/pvz/datadots/presentation/results/ResultsFragment.kt
@@ -4,18 +4,25 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.asLiveData
+import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.github.mikephil.charting.components.XAxis
 import com.github.mikephil.charting.data.Entry
 import com.github.mikephil.charting.data.LineData
 import com.github.mikephil.charting.data.LineDataSet
+import com.pvz.datadots.R
 import com.pvz.datadots.databinding.FragmentResultsBinding
 import com.pvz.datadots.domain.model.Point
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class ResultsFragment : Fragment() {
@@ -31,17 +38,50 @@ class ResultsFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         viewModel.fetchPointsData()
+
         viewModel.pointList.observe(viewLifecycleOwner) { points ->
             if (points.isEmpty()) {
-                Toast.makeText(requireContext(), "Нет данных для отображения", Toast.LENGTH_SHORT).show()
+                Toast.makeText(requireContext(), "Нет данных для отображения", Toast.LENGTH_SHORT)
+                    .show()
                 findNavController().popBackStack()
             }
             with(binding.pointsRecyclerView) {
                 layoutManager = LinearLayoutManager(context)
                 adapter = PointsAdapter(points)
             }
-
             setupChart(points)
+            viewModel.spinnerId.asLiveData().observe(viewLifecycleOwner) { position ->
+                when (position) {
+                    0 -> setLineChartMode(LineDataSet.Mode.LINEAR)
+                    1 -> setLineChartMode(LineDataSet.Mode.CUBIC_BEZIER)
+                    else -> setLineChartMode(LineDataSet.Mode.LINEAR)
+                }
+            }
+        }
+        with(binding.spinner) {
+            val smoothingOptions =
+                requireContext().resources.getStringArray(R.array.dropdown_items)
+            val adapter = ArrayAdapter(
+                requireContext(),
+                android.R.layout.simple_spinner_item,
+                smoothingOptions
+            )
+            adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+            this.adapter = adapter
+            onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+                override fun onItemSelected(
+                    parentView: AdapterView<*>,
+                    selectedItemView: View?,
+                    position: Int,
+                    id: Long
+                ) {
+                    viewModel.updateSpinnerItemId(position)
+                }
+
+                override fun onNothingSelected(parentView: AdapterView<*>) {
+                    // Do nothing
+                }
+            }
         }
 
         return binding.root
@@ -61,7 +101,7 @@ class ResultsFragment : Fragment() {
             mode = LineDataSet.Mode.CUBIC_BEZIER
         }
 
-        with(binding.lineChart){
+        with(binding.lineChart) {
             data = LineData(dataSet)
 
             xAxis.apply {
@@ -76,8 +116,19 @@ class ResultsFragment : Fragment() {
         }
     }
 
+    private fun setLineChartMode(mode: LineDataSet.Mode) {
+        val data = binding.lineChart.data
+        if (data != null && data.dataSets.isNotEmpty()) {
+            val dataSet = data.getDataSetByIndex(0) as LineDataSet
+            dataSet.mode = mode
+            binding.lineChart.invalidate()
+        }
+    }
+
     override fun onDestroyView() {
         super.onDestroyView()
-        viewModel.deleteAllPoints()
+        if (isRemoving) {
+            viewModel.deleteAllPoints()
+        }
     }
 }
